+++ src/io/tpr_io.rs
@@ -12,9 +12,10 @@ use crate::{
 };
 
 /// Read topology from a tpr file. Construct a system.
-/// Does not read intermolecular bonds, does not read positions/velocities/forces,
-/// does not create groups (other than the default `all` and `All`).
-/// Note that the atoms and residues in the tpr file are numbered sequentially,
+///
+/// ## Notes
+/// - Does not create groups apart from the default `all` and `All`.
+/// - The atoms and residues in the tpr file are numbered sequentially,
 /// no matter their original numbering in the gro/pdb file.
 pub fn read_tpr(filename: impl AsRef<Path>) -> Result<System, ParseTprError> {
     let tpr = match minitpr::TprFile::parse(filename) {
@@ -45,7 +46,7 @@ pub fn read_tpr(filename: impl AsRef<Path>) -> Result<System, ParseTprError> {
 impl From<minitpr::Atom> for Atom {
     /// Convert `Atom` from `minitpr` crate to native `groan_rs` Atom.
     fn from(value: minitpr::Atom) -> Self {
-        let atom = Self::new(
+        let mut atom = Self::new(
             value.residue_number as usize,
             &value.residue_name,
             value.atom_number as usize,
@@ -54,12 +55,24 @@ impl From<minitpr::Atom> for Atom {
         .with_mass(value.mass as f32)
         .with_charge(value.charge as f32);
 
-        match value.element {
-            None => atom,
-            Some(x) => atom
-                .with_element_name(&x.name().to_string().to_lowercase())
-                .with_element_symbol(x.symbol()),
+        if let Some(pos) = value.position {
+            atom.set_position(pos.into())
         }
+
+        if let Some(vel) = value.velocity {
+            atom.set_velocity(vel.into())
+        }
+
+        if let Some(force) = value.force {
+            atom.set_force(force.into())
+        }
+
+        if let Some(element) = value.element {
+            atom.set_element_name(&element.name().to_string().to_lowercase());
+            atom.set_element_symbol(element.symbol());
+        }
+
+        atom
     }
 }
 
@@ -84,6 +97,8 @@ impl From<minitpr::SimBox> for SimBox {
 mod tests {
     use float_cmp::assert_approx_eq;
 
+    use crate::{io::gro_io::read_gro, structures::vector3d::Vector3D};
+
     use super::*;
 
     #[test]
@@ -96,6 +111,9 @@ mod tests {
             mass: 72.0,
             charge: 1.0,
             element: Some(minitpr::Element::C),
+            position: Some([1.497, 2.432, 0.439]),
+            velocity: None,
+            force: Some([17.3424, -7.2974, 0.3412]),
         };
 
         let atom2 = minitpr::Atom {
@@ -106,6 +124,9 @@ mod tests {
             mass: 36.0,
             charge: 0.0,
             element: None,
+            position: None,
+            velocity: Some([-1.424, 0.439, -2.432]),
+            force: None,
         };
 
         let converted1: Atom = atom1.into();
@@ -119,9 +140,15 @@ mod tests {
         assert_approx_eq!(f32, converted1.get_charge().unwrap(), 1.0);
         assert_eq!(converted1.get_element_name(), Some("carbon"));
         assert_eq!(converted1.get_element_symbol(), Some("C"));
-        assert!(!converted1.has_position());
+        assert_eq!(
+            converted1.get_position(),
+            Some(&Vector3D::new(1.497, 2.432, 0.439))
+        );
         assert!(!converted1.has_velocity());
-        assert!(!converted1.has_force());
+        assert_eq!(
+            converted1.get_force(),
+            Some(&Vector3D::new(17.3424, -7.2974, 0.3412))
+        );
 
         assert_eq!(converted2.get_atom_name(), "SC1");
         assert_eq!(converted2.get_atom_number(), 2);
@@ -132,7 +159,10 @@ mod tests {
         assert_eq!(converted2.get_element_name(), None);
         assert_eq!(converted2.get_element_symbol(), None);
         assert!(!converted2.has_position());
-        assert!(!converted2.has_velocity());
+        assert_eq!(
+            converted2.get_velocity(),
+            Some(&Vector3D::new(-1.424, 0.439, -2.432))
+        );
         assert!(!converted2.has_force());
     }
 
@@ -163,46 +193,29 @@ mod tests {
     #[test]
     fn parse_tpr() {
         let system_tpr = read_tpr("test_files/example.tpr").unwrap();
+        let system_gro = read_gro("test_files/example.gro").unwrap();
 
-        assert_eq!(system_tpr.get_n_atoms(), 16844);
+        assert_eq!(system_tpr.get_n_atoms(), system_gro.get_n_atoms());
 
-        let atom1 = system_tpr.get_atom_as_ref(0).unwrap();
-        let atom2 = system_tpr.get_atom_as_ref(16843).unwrap();
+        for (a1, a2) in system_tpr.atoms_iter().zip(system_gro.atoms_iter()) {
+            crate::test_utilities::utilities::compare_atoms_tpr_with_gro(a1, a2);
+            assert!(!a1.has_force());
+        }
 
-        assert_eq!(atom1.get_atom_name(), "BB");
-        assert_eq!(atom1.get_atom_number(), 1);
-        assert_eq!(atom1.get_residue_name(), "GLY");
-        assert_eq!(atom1.get_residue_number(), 1);
-        assert_approx_eq!(f32, atom1.get_charge().unwrap(), 1.0);
-        assert_approx_eq!(f32, atom1.get_mass().unwrap(), 72.0);
-        assert!(atom1.get_element_name().is_none());
-        assert!(atom1.get_element_symbol().is_none());
-        assert!(!atom1.has_position());
-        assert!(!atom1.has_velocity());
-        assert!(!atom1.has_force());
-
-        assert_eq!(atom2.get_atom_name(), "CL");
-        assert_eq!(atom2.get_atom_number(), 16844);
-        assert_eq!(atom2.get_residue_name(), "ION");
-        assert_eq!(atom2.get_residue_number(), 11180);
-        assert_approx_eq!(f32, atom2.get_charge().unwrap(), -1.0);
-        assert_approx_eq!(f32, atom2.get_mass().unwrap(), 35.453);
-        assert!(atom2.get_element_name().is_none());
-        assert!(atom2.get_element_symbol().is_none());
-        assert!(!atom2.has_position());
-        assert!(!atom2.has_velocity());
-        assert!(!atom2.has_force());
-
-        let bonded1 = atom1.get_bonded();
-        assert_eq!(bonded1.get_n_atoms(), 1);
-        assert!(bonded1.isin(1));
-
-        let bonded2 = atom2.get_bonded();
-        assert!(bonded2.is_empty());
-
-        let bonded_special = system_tpr.get_atom_as_ref(2).unwrap().get_bonded();
-        assert_eq!(bonded_special.get_n_atoms(), 2);
-        assert!(bonded_special.isin(1));
-        assert!(bonded_special.isin(3));
+        // check bonds
+        system_tpr.get_atom_as_ref(0).unwrap().get_bonded().isin(1);
+        system_tpr.get_atom_as_ref(1).unwrap().get_bonded().isin(0);
+        system_tpr.get_atom_as_ref(1).unwrap().get_bonded().isin(2);
+        system_tpr.get_atom_as_ref(1).unwrap().get_bonded().isin(4);
+        system_tpr
+            .get_atom_as_ref(854)
+            .unwrap()
+            .get_bonded()
+            .isin(853);
+        system_tpr
+            .get_atom_as_ref(854)
+            .unwrap()
+            .get_bonded()
+            .isin(855);
     }
 }