+++ changelog.md
@@ -3,6 +3,9 @@
 
 ### Version 0.8.0
 
+#### Reading structures from TPR files
+- In addition to system topology and simulation box, the positions, velocities, and forces of atoms and intermolecular bonds can be now also read from TPR files.
+
 #### Creating Atom iterators directly from selection queries
 - Iterators over atoms of the system can be now constructed directly from selection queries using `System::selection_iter` (for iteration over immutable atoms) and using `System::selection_iter_mut` (for iteration over mutable atoms).
 - This allows selecting atoms without adding groups into the system.